@@ -2,14 +2,14 @@ use std::collections::HashMap;
 use std::path::{Path, PathBuf};
 use std::sync::Arc;
 
-use anyhow::{anyhow, bail, Context, Result};
+use anyhow::{Context, Result, anyhow, bail};
 use async_recursion::async_recursion;
 use async_trait::async_trait;
 use clap::Parser;
 use serde::Deserialize;
 use tokio::fs;
 use tokio::sync::Mutex;
-use tracing::{event, info, instrument, warn, Level};
+use tracing::{Level, event, info, instrument, warn};
 use uuid::Uuid;
 
 use treadmill_rs::api::switchboard_supervisor::{
@@ -83,16 +83,12 @@ pub struct NbdNetbootSupervisorArgs {
 
 #[derive(Deserialize, Debug, Clone)]
 #[serde(rename_all = "snake_case")]
+#[derive(Default)]
 pub enum SSHPreferredIPVersion {
+    #[default]
     Unspecified,
     V4,
     V6,
-}
-
-impl Default for SSHPreferredIPVersion {
-    fn default() -> Self {
-        SSHPreferredIPVersion::Unspecified
-    }
 }
 
 #[derive(Deserialize, Debug, Clone)]
@@ -369,7 +365,11 @@ impl NbdNetbootSupervisor {
                         // We could not retrieve the manifest, despite the image
                         // being present. Don't attempt to recover from this
                         // error and mark the job as failed:
-                        event!(Level::WARN, ?manifest_err, "Failed to retrieve image manifest, reporting job error: {manifest_err:?}");
+                        event!(
+                            Level::WARN,
+                            ?manifest_err,
+                            "Failed to retrieve image manifest, reporting job error: {manifest_err:?}"
+                        );
                         this.connector
                             .report_job_error(
                                 fetching_image_state.start_job_req.job_id,
@@ -509,7 +509,7 @@ impl NbdNetbootSupervisor {
             .ok_or(anyhow!("Image does not have a head layer defined"))?;
 
         // Check that we have a boot archive defined:
-        let boot_blob_label = format!("{}-boot", top_layer_label);
+        let boot_blob_label = format!("{top_layer_label}-boot");
         let boot_blob_spec = manifest
             .blobs
             .get(&boot_blob_label)
@@ -523,9 +523,7 @@ impl NbdNetbootSupervisor {
 
         let boot_blob_path_canon = tokio::fs::canonicalize(&boot_blob_path)
             .await
-            .with_context(|| {
-                format!("Failed to canonicalize boot blob path {:?}", boot_blob_path)
-            })?;
+            .with_context(|| format!("Failed to canonicalize boot blob path {boot_blob_path:?}"))?;
 
         // Now step through each image layer, making sure that it has a root
         // disk and (if that has a predecessor defined) the current root blob
@@ -541,7 +539,7 @@ impl NbdNetbootSupervisor {
         // - Matches its specification in the manifest, and
         // - Only references its defined predecessor.
         let mut top_root_image = None;
-        let mut current_root_image = format!("{}-root", top_layer_label);
+        let mut current_root_image = format!("{top_layer_label}-root");
         loop {
             // Acquire the manifest metadata for this image blob:
             let blob_spec = manifest
@@ -559,7 +557,7 @@ impl NbdNetbootSupervisor {
                 tokio::fs::canonicalize(&image_path)
                     .await
                     .with_context(|| {
-                        format!("Failed to canonicalize image blob path {:?}", image_path)
+                        format!("Failed to canonicalize image blob path {image_path:?}")
                     })?;
 
             event!(Level::DEBUG, qemu_img_binary = ?self.config.nbd_netboot.qemu_img_binary, file = ?image_path_canon, "Retrieving qcow2 file metadata");
@@ -572,7 +570,7 @@ impl NbdNetbootSupervisor {
                     .arg(&image_path_canon)
                     .output()
                     .await
-                    .map_err(|e| anyhow::Error::from(e))
+                    .map_err(anyhow::Error::from)
                     .and_then(|output| {
                         // Ideally we'd want to use the nightly `exit_ok()` here:
                         if !output.status.success() {
@@ -586,7 +584,7 @@ impl NbdNetbootSupervisor {
                         Ok(output.stdout)
                     })
                     .with_context(|| {
-                        format!("Failed to query image metadata for {:?}", image_path)
+                        format!("Failed to query image metadata for {image_path:?}")
                     })?;
 
             let metadata: QemuImgMetadata =
@@ -606,7 +604,7 @@ impl NbdNetbootSupervisor {
             // Make sure that the image has just one child, and that child's
             // filename is identical to the top-level filename. We do this as a
             // precaution to images referencing other paths on our machine.
-            if metadata.children.len() != 1 || metadata.children[0].info.children.len() != 0 {
+            if metadata.children.len() != 1 || !metadata.children[0].info.children.is_empty() {
                 bail!(
                     "Image file {:?} does not have exactly one (recursive) child in qemu-img output",
                     image_path
@@ -638,8 +636,7 @@ impl NbdNetbootSupervisor {
                 .and_then(|size_bytes_str| {
                     u64::from_str_radix(size_bytes_str, 10).with_context(|| {
                         format!(
-                            "Parsing blob {}'s `blob-virtual-size` attribute as u64",
-                            current_root_image
+                            "Parsing blob {current_root_image}'s `blob-virtual-size` attribute as u64"
                         )
                     })
                 })?;
@@ -700,18 +697,16 @@ impl NbdNetbootSupervisor {
                     .await
                     .with_context(|| {
                         format!(
-                            "Failed to canonicalize image blob path {:?},
-                             returned by image store lookup for blob {}",
-                            image_store_path, blob_lower,
+                            "Failed to canonicalize image blob path {image_store_path:?},
+                             returned by image store lookup for blob {blob_lower}",
                         )
                     })?;
 
                 let full_backing_filename_canon = if let Some(f) = &metadata.full_backing_filename {
                     tokio::fs::canonicalize(f).await.with_context(|| {
                         format!(
-                            "Failed to canonicalize image blob path {:?},
-                                     referenced by qcow image file {}",
-                            f, current_root_image,
+                            "Failed to canonicalize image blob path {f:?},
+                                     referenced by qcow image file {current_root_image}",
                         )
                     })?
                 } else {
@@ -785,8 +780,7 @@ impl NbdNetbootSupervisor {
             .map_err(|io_err| connector::JobError {
                 error_kind: connector::JobErrorKind::InternalError,
                 description: format!(
-                    "Unable to create (parents of) TFTP boot directory at {:?}: {:?}",
-                    tftp_boot_dir, io_err
+                    "Unable to create (parents of) TFTP boot directory at {tftp_boot_dir:?}: {io_err:?}"
                 ),
             })?;
 
@@ -820,8 +814,7 @@ impl NbdNetbootSupervisor {
             .map_err(|io_err| connector::JobError {
                 error_kind: connector::JobErrorKind::InternalError,
                 description: format!(
-                    "Unable to delete files of TFTP boot directory at {:?}: {:?}",
-                    tftp_boot_dir, io_err
+                    "Unable to delete files of TFTP boot directory at {tftp_boot_dir:?}: {io_err:?}"
                 ),
             })?;
 
@@ -840,13 +833,13 @@ impl NbdNetbootSupervisor {
             .arg("-x")
             // Target directory:
             .arg("-C")
-	    .arg(&tftp_boot_dir)
+	    .arg(tftp_boot_dir)
 	    // Boot archive:
             .arg("-f")
-            .arg(&boot_archive_path)
+            .arg(boot_archive_path)
             .output()
             .await
-            .map_err(|e| anyhow::Error::from(e))
+            .map_err(anyhow::Error::from)
             .and_then(|output| {
                 // Ideally we'd want to use the nightly `exit_ok()` here:
                 if !output.status.success() {
@@ -880,7 +873,7 @@ impl NbdNetbootSupervisor {
         top_root_image_qemu_info: &QemuImgMetadata,
     ) -> Result<(PathBuf, PathBuf), connector::JobError> {
         let jobs_dir = self.config.nbd_netboot.state_dir.join("jobs");
-        let job_dir = jobs_dir.join(&start_job_req.job_id.to_string());
+        let job_dir = jobs_dir.join(start_job_req.job_id.to_string());
 
         // Ensure that the state/jobs directory (and all recursive parents)
         // exists and create a new working directory for this job:
@@ -961,14 +954,14 @@ impl NbdNetbootSupervisor {
             .arg("qcow2")
             // Backing file:
             .arg("-b")
-            .arg(&top_root_image_path)
+            .arg(top_root_image_path)
             // New image file:
             .arg(&disk_image_file)
             // New image size (in bytes, without suffix):
-            .arg(&self.config.nbd_netboot.working_disk_max_bytes.to_string())
+            .arg(self.config.nbd_netboot.working_disk_max_bytes.to_string())
             .output()
             .await
-            .map_err(|e| anyhow::Error::from(e))
+            .map_err(anyhow::Error::from)
             .and_then(|output| {
                 // Ideally we'd want to use the nightly `exit_ok()` here:
                 if !output.status.success() {
@@ -1065,38 +1058,37 @@ impl NbdNetbootSupervisor {
 
         // Parse the manifest and sanity-check all image layers. We do this in
         // an async block such that we can use the ``?`` operator:
-        let (boot_archive_path, top_root_image_path, top_root_image_qemu_info) =
-            match this.start_job_parse_manifest_check_images(&manifest).await {
-                Ok(v) => v,
+        let (boot_archive_path, top_root_image_path, top_root_image_qemu_info) = match this
+            .start_job_parse_manifest_check_images(&manifest)
+            .await
+        {
+            Ok(v) => v,
 
-                Err(e) => {
-                    // Image is invalid, report an error:
-                    this.connector
-                        .report_job_error(
-                            start_job_req.job_id,
-                            connector::JobError {
-                                error_kind: connector::JobErrorKind::ImageInvalid,
-                                description: format!(
-                                    "Validation of image {:?} failed: {:?}",
-                                    image_id, e
-                                ),
-                            },
-                        )
-                        .await;
+            Err(e) => {
+                // Image is invalid, report an error:
+                this.connector
+                    .report_job_error(
+                        start_job_req.job_id,
+                        connector::JobError {
+                            error_kind: connector::JobErrorKind::ImageInvalid,
+                            description: format!("Validation of image {image_id:?} failed: {e:?}"),
+                        },
+                    )
+                    .await;
 
-                    // No resources have been allocated (yet), so simply remove the
-                    // job and set its state to `Stopping`, in case anyone else has
-                    // a reference to it still.
-                    //
-                    // Safe to call, we don't hold a lock on `this.jobs`:
-                    this.remove_job(job_id).await;
+                // No resources have been allocated (yet), so simply remove the
+                // job and set its state to `Stopping`, in case anyone else has
+                // a reference to it still.
+                //
+                // Safe to call, we don't hold a lock on `this.jobs`:
+                this.remove_job(job_id).await;
 
-                    // Prevent other tasks from further advancing this job's state:
-                    *job_lg = NbdNetbootSupervisorJobState::Stopping;
+                // Prevent other tasks from further advancing this job's state:
+                *job_lg = NbdNetbootSupervisorJobState::Stopping;
 
-                    return;
-                }
-            };
+                return;
+            }
+        };
 
         // Unpack the boot archive:
         if let Err(job_error) = this
@@ -1200,7 +1192,7 @@ impl NbdNetbootSupervisor {
 	    .current_dir(&job_workdir)
             .output()
             .await
-            .map_err(|e| anyhow::Error::from(e))
+            .map_err(anyhow::Error::from)
             .and_then(|output| {
                 // Ideally we'd want to use the nightly `exit_ok()` here:
                 if !output.status.success() {
@@ -1317,7 +1309,7 @@ impl NbdNetbootSupervisor {
                     // process:
                     if let Err(e) = qemu_proc.kill().await {
                         // We were unable to terminate the QEMU process:
-                        let message = format!("Failed to wait on qemu-nbd process: {:?}", e);
+                        let message = format!("Failed to wait on qemu-nbd process: {e:?}");
 
                         this.connector.report_job_error(
                             job_id,
@@ -1349,7 +1341,7 @@ impl NbdNetbootSupervisor {
                                 "qemu-nbd process exited successfully.".to_string()
                             } else {
                                 // Notify failure
-                                let message = format!("qemu-nbd process had an internal error with status: {:?}", status);
+                                let message = format!("qemu-nbd process had an internal error with status: {status:?}");
 
                                 this.connector.report_job_error(
                                     job_id,
@@ -1365,7 +1357,7 @@ impl NbdNetbootSupervisor {
                         }
                         Err(e) => {
                             // Failed to wait on qemu-nbd process
-                            let message = format!("Failed to wait on qemu-nbd process: {:?}", e);
+                            let message = format!("Failed to wait on qemu-nbd process: {e:?}");
 
                             this.connector.report_job_error(
                                 job_id,
@@ -1443,7 +1435,7 @@ impl NbdNetbootSupervisor {
                 .cloned()
                 .ok_or(connector::JobError {
                     error_kind: connector::JobErrorKind::JobNotFound,
-                    description: format!("Job {:?} not found, cannot stop.", job_id),
+                    description: format!("Job {job_id:?} not found, cannot stop."),
                 })?
         };
 
@@ -1485,7 +1477,7 @@ impl NbdNetbootSupervisor {
 
                     return Err(connector::JobError {
                         error_kind: connector::JobErrorKind::AlreadyStopping,
-                        description: format!("Job {:?} is already stopping.", job_id),
+                        description: format!("Job {job_id:?} is already stopping."),
                     });
                 }
             };
@@ -1584,7 +1576,7 @@ impl NbdNetbootSupervisor {
 	    .current_dir(&job_workdir)
             .output()
             .await
-            .map_err(|e| anyhow::Error::from(e))
+            .map_err(anyhow::Error::from)
             .and_then(|output| {
                 // Ideally we'd want to use the nightly `exit_ok()` here:
                 if !output.status.success() {
@@ -1801,7 +1793,7 @@ impl control_socket::Supervisor for NbdNetbootSupervisor {
             Some(job_state) => match &*job_state.lock().await {
                 // Job is currently running, respond with its assigned hostname:
                 NbdNetbootSupervisorJobState::Running(_) => {
-                    let hostname = format!("job-{}", format!("{}", tgt_job_id).split_at(10).0);
+                    let hostname = format!("job-{}", format!("{tgt_job_id}").split_at(10).0);
                     Some(treadmill_rs::api::supervisor_puppet::NetworkConfig {
                         hostname,
                         // NbdNetbootSupervisor, don't supply a network interface to configure:
@@ -2000,7 +1992,7 @@ async fn main() -> Result<()> {
 
     match config.base.coord_connector {
         SupervisorCoordConnector::WsConnector => {
-            use tokio::signal::unix::{signal, SignalKind};
+            use tokio::signal::unix::{SignalKind, signal};
 
             let ws_connector_config = config.ws_connector.clone().ok_or(anyhow!(
                 "Requested WsConnector, but `ws_connector` config not present."
