@@ -5,8 +5,8 @@ use crate::sql;
 use crate::sql::api_token::TokenPerms;
 use argon2::password_hash::{PasswordHashString, SaltString};
 use argon2::{Algorithm, Argon2, Params, PasswordHasher, PasswordVerifier, Version};
-use axum::extract::{ConnectInfo, State};
 use axum::Json;
+use axum::extract::{ConnectInfo, State};
 use http::StatusCode;
 use sqlx::{Error, PgExecutor};
 use std::net::SocketAddr;
