@@ -1,8 +1,8 @@
 use crate::config::{DatabaseConfig, DatabaseCredentials, SwitchboardConfig};
 use crate::service::Service;
 use miette::{IntoDiagnostic, WrapErr};
-use sqlx::postgres::PgConnectOptions;
 use sqlx::PgPool;
+use sqlx::postgres::PgConnectOptions;
 use std::net::SocketAddr;
 use std::ops::Deref;
 use std::path::PathBuf;
@@ -59,8 +59,7 @@ pub async fn pg_pool_from_config(db_config: &DatabaseConfig) -> Result<PgPool, s
 }
 
 pub async fn serve(serve_command: ServeCommand) -> miette::Result<()> {
-    let config =
-        super::config::load_configuration(serve_command.config.as_ref().map(PathBuf::as_path))?;
+    let config = super::config::load_configuration(serve_command.config.as_deref())?;
 
     if config.log.use_tokio_console_subscriber {
         console_subscriber::init();
@@ -112,7 +111,9 @@ pub async fn serve(serve_command: ServeCommand) -> miette::Result<()> {
                     .wrap_err("Failed to load RusTls configuration for public server")?;
             let server = axum_server::bind_rustls(bind_address, rustls_config);
 
-            tracing::warn!("-- WARNING -- DEVELOPMENT-ONLY TLS MODE IS ENABLED. PLEASE DO NOT USE THIS IN PRODUCTION.");
+            tracing::warn!(
+                "-- WARNING -- DEVELOPMENT-ONLY TLS MODE IS ENABLED. PLEASE DO NOT USE THIS IN PRODUCTION."
+            );
 
             Server::Tls(server)
         }
