@@ -20,8 +20,8 @@ use uuid::Uuid;
 /// Returns `(join_handle, outbox_send, event_recv)`.
 /// - `join_handle` is a future that completes when the connection with the supervisor closes.
 /// - `outbox_send` is used for sending messages, and takes values of the form
-/// `(mesg, Option<resp_send>)`, where the optional one-shot response channel can be used to
-/// indicate that a response is expected, and pass the response back to the caller.
+///   `(mesg, Option<resp_send>)`, where the optional one-shot response channel can be used to
+///   indicate that a response is expected, and pass the response back to the caller.
 /// - `event_recv` is used for receiving messages sent unilaterally from the supervisor.
 pub fn supervisor_run_loop(
     supervisor_id: Uuid,
@@ -195,7 +195,7 @@ impl SupervisorConnection {
                 if let Some((_, notifier)) =
                     self.outstanding_requests.remove(&rm.response_to_request_id)
                 {
-                    if let Err(_) = notifier.send(rm.message) {
+                    if notifier.send(rm.message).is_err() {
                         tracing::error!(
                             "failed to send response message via outstanding channel: receiver dropped"
                         );
@@ -240,7 +240,9 @@ impl SupervisorConnection {
                     let m: switchboard_supervisor::Message = match serde_json::from_str(&s) {
                         Ok(m) => m,
                         Err(e) => {
-                            tracing::error!("Error deserializing message ({s}) from supervisor ({supervisor_id}): {e}");
+                            tracing::error!(
+                                "Error deserializing message ({s}) from supervisor ({supervisor_id}): {e}"
+                            );
                             self.try_close(socket, supervisor_id, None).await;
                             return ControlFlow::Break(());
                         }
